The Newtonator is an FSU distortion unit
The Newtonator is an FSU distortion unit of my own design that is more than capable of mangling the input audio beyond recognition. It is also capable of more subtle distortion, and can act as a low-pass filter with the correct settings. Please read the comments for instructions.
Original author: Michael Bechard, Jan. 2008